requirements.txt

Django==5.2.1
djangorestframework==3.16.0
djongo5==1.3.9
pymongo==4.12.1
dnspython==2.7.0
pillow==11.2.1
django-tailwind==3.8.0

tensorflow==2.19.0; 
tf-nightly==2.20.0.dev20250508; 
opencv-python-headless==4.11.0.86

python-dotenv==1.1.0

# task queue / cache
celery==5.5.2
redis==6.0.0  

# auth & CORS
djoser==2.3.1
django-cors-headers==4.7.0
